refactor(QUAL-49): one serializer per entity (persisted-but-dropped class)
Extract per-entity serializers into src/backend/serializers/ (city, place,
trip, item) and route every trip/place/city/item-emitting handler through
them. Collapses the duplicated city hand-maps and the two trips.ts place
hand-maps into shared, composed serializers — region_iso/region_name (BUG-80)
and place-date (BUG-31) fields are now structural, not hand-copied.

Response shapes preserved byte-for-byte (16 characterization tests + 827
existing green). The item serializer is flattenItem moved verbatim out of
items-helper.ts; the query helper now imports it.
